backend/Inventories/delete-inventories/requirements.txt
flask
flask-cors
flask-jwt-extended
pymysql
python-dotenv
